Luna branch

+++ public/js/members.js
@@ -8,6 +8,8 @@ document.addEventListener('DOMContentLoaded', (e) => {
   const todoContainer = document.querySelector('.todo-container');
   const todoForm = document.getElementById('todo-form');
 
+  let userID = 0;
+
   // Inital todos array
   let todos = [];
 
@@ -32,6 +34,18 @@ document.addEventListener('DOMContentLoaded', (e) => {
 
   // Helper function to grab todos
   const getTodos = () => {
+    fetch('/api/user_data', {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        console.log('user:', data.id);
+        userID = data.id;
+      });
+
     fetch('/api/todos', {
       method: 'GET',
       headers: {
@@ -52,32 +66,63 @@ document.addEventListener('DOMContentLoaded', (e) => {
   const deleteTodo = (e) => {
     e.stopPropagation();
     const { id } = e.target.dataset;
-
     fetch(`/api/todos/${id}`, {
-      method: 'DELETE',
+      method: 'GET',
       headers: {
         'Content-Type': 'application/json',
       },
-    }).then(getTodos);
+    }).then((response) => response.json())
+    .then((data) => {
+      console.log('todo by id:', data[0]);
+      if (userID == data[0].userId) {
+        fetch(`/api/todos/${id}`, {
+          method: 'DELETE',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        }).then(getTodos);
+      } else {
+        alert("not your post!!!");
+      }
+    });
+
+    
   };
 
   // Function to handle the editing of a todo when input is clicked
   const editTodo = (e) => {
-    const itemChildren = e.target.children;
-    // console.log('editTodo -> itemChildren', itemChildren);
-    for (let i = 0; i < itemChildren.length; i++) {
-      const currentEl = itemChildren[i];
-
-      if (currentEl.tagName === 'INPUT') {
-        currentEl.value = currentEl.parentElement.children[0].innerText;
-        show(currentEl);
-        currentEl.focus();
+    e.stopPropagation();
+    const { id } = e.target.dataset;
+    console.log(id)
+    fetch(`/api/todos/${id}`, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    }).then((response) => response.json())
+    .then((data) => {
+      console.log('todo by id:', data[0]);
+      if (userID == data[0].userId) {
+        const itemChildren = e.target.children;
+        // console.log('editTodo -> itemChildren', itemChildren);
+        for (let i = 0; i < itemChildren.length; i++) {
+          const currentEl = itemChildren[i];
+    
+          if (currentEl.tagName === 'INPUT') {
+            currentEl.value = currentEl.parentElement.children[0].innerText;
+            show(currentEl);
+            currentEl.focus();
+          }
+    
+          if (currentEl.tagName === 'SPAN' || currentEl.tagName === 'BUTTON') {
+            hide(currentEl);
+          }
+        }
+      } else {
+        alert("You don't have access, Your input was illegal and has been reported to the athurities");
       }
+    });
 
-      if (currentEl.tagName === 'SPAN' || currentEl.tagName === 'BUTTON') {
-        hide(currentEl);
-      }
-    }
   };
 
   // Function to handle when a user cancels editing
@@ -144,16 +189,20 @@ document.addEventListener('DOMContentLoaded', (e) => {
     const newInputRow = document.createElement('li');
     newInputRow.classList.add('list-group-item', 'todo-item');
     newInputRow.setAttribute('complete', todo.complete);
+    newInputRow.setAttribute('data-id', todo.id);
 
     // Span
     const rowSpan = document.createElement('span');
     rowSpan.innerText = todo.text;
+    rowSpan.setAttribute('data-id', todo.id);
 
     // Input field
     const rowInput = document.createElement('input');
     rowInput.setAttribute('type', 'text');
     rowInput.classList.add('edit');
     rowInput.style.display = 'none';
+    rowInput.setAttribute('data-id', todo.id);
+
 
     // Delete button
     const delBtn = document.createElement('button');
@@ -173,6 +222,7 @@ document.addEventListener('DOMContentLoaded', (e) => {
 
     // Add event listener for editing
     newInputRow.addEventListener('click', editTodo);
+    newInputRow.setAttribute('data-id', todo.id);
     rowInput.addEventListener('blur', cancelEdit);
     rowInput.addEventListener('keyup', finishEdit);
 
@@ -205,11 +255,15 @@ document.addEventListener('DOMContentLoaded', (e) => {
 
   // Function to actually put the todo on the page
   const insertTodo = (e) => {
+    console.log("TESTING: " + userID);
     e.preventDefault();
     const todo = {
       text: document.getElementById('newTodo').value.trim(),
       complete: false,
+      userId: userID
     };
+
+    console.log("Test2: " + JSON.stringify(todo));
     if (todo.text) {
       fetch('/api/todos', {
         method: 'POST',
